Fixed minor nil pointer exception
Added more appengine support in that appengine/mail.Message can be sent using the sendgrid library.  This makes it a very minimal change to code to move to sendgrid.
appengine support requires storing the sendgrid user/pass information in the datastore.  Messages are automatically queued as tasks using appengine/delay