Remove mail.send helpers with on-behalf-of header
The sendgrid documentation (and my own findings) say that the mail.send endpoint
does not support using the `on-behalf-of` header.

[Documentation](
https://docs.sendgrid.com/api-reference/how-to-use-the-sendgrid-v3-api/on-behalf-of-subuser
)

Providing the `NewSendClientSubuser` in this SDK causes wasted time because it
sets the expectation that this is possible when it is not.

I also added guards to the `GetRequestSubuser` function to prevent people from
using it with the mail.send endpoint.